requirements.txt
attrs==20.3.0
jsonschema==3.2.0
pyrsistent==0.17.3
PyYAML==5.4.1
six==1.15.0
